Move .ts sources from / into /src/
