+++ src/index.ts
@@ -17,13 +17,13 @@
 import * as fs from 'fs';
 import * as postcss from 'postcss';
 import * as selectorParser from 'postcss-selector-parser';
-import { IdentitySubstitutionMap } from './src/com/google/common/css/identity-substitution-map';
-import { MinimalSubstitutionMap } from './src/com/google/common/css/minimal-substitution-map';
-import { OutputRenamingMapFormat } from './src/com/google/common/css/output-renaming-map-format';
-import { PrefixingSubstitutionMap } from './src/com/google/common/css/prefixing-substitution-map';
-import { RecordingSubstitutionMap } from './src/com/google/common/css/recording-substitution-map';
-import { SimpleSubstitutionMap } from './src/com/google/common/css/simple-substitution-map';
-import { SplittingSubstitutionMap } from './src/com/google/common/css/splitting-substitution-map';
+import { IdentitySubstitutionMap } from './com/google/common/css/identity-substitution-map';
+import { MinimalSubstitutionMap } from './com/google/common/css/minimal-substitution-map';
+import { OutputRenamingMapFormat } from './com/google/common/css/output-renaming-map-format';
+import { PrefixingSubstitutionMap } from './com/google/common/css/prefixing-substitution-map';
+import { RecordingSubstitutionMap } from './com/google/common/css/recording-substitution-map';
+import { SimpleSubstitutionMap } from './com/google/common/css/simple-substitution-map';
+import { SplittingSubstitutionMap } from './com/google/common/css/splitting-substitution-map';
 
 const RENAMING_TYPE = {
   none: () => new IdentitySubstitutionMap(),
