+++ tsconfig.json
@@ -20,6 +20,7 @@
     "allowJs": true
   },
   "include": [
-    "*.ts", "src/com/google/common/css/*.js"
+    "src/com/google/common/css/*.js",
+    "src/index.ts"
   ],
 }
